backend/requirements.txt
Flask
Flask-CORS
pyjwt
python-dotenv
bcrypt